src/app/requirements.txt
Django==4.1.4
drf_yasg==1.21.4
django_filter==22.1
django-extensions==3.2.1
django-rest-swagger==2.2.0
django_cors_headers==3.13.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.1
django_phonenumber_field==7.0.0
django_phonenumbers==1.0.1
django_timezone_field===5.0
django_constance===2.9.1
django-picklefield===3.1
celery==5.2.7
cleo==1.0.0a5
certifi==2022.9.24
Pillow==9.3.0
redis==4.4.0
async-timeout==4.0.2
django-db-logger==0.1.12
ipython==8.7.0
autopep8==2.0.0
bcrypt==4.0.1